reference_implementations/gcp/vertex/predictor/requirements.txt
# requirements for the custom predictor, used by build_and_push_image.py

google-cloud-aiplatform[prediction]==1.66.0
torch==2.2.0
transformers==4.44.0
accelerate==0.27.0
pandas==2.0.3
google-cloud-bigquery==3.26.0
joblib==1.4.2
numpy<2
